Handle duplicate filenames

+++ pawoo_dl.py
@@ -1,5 +1,9 @@
 import os, sys, urllib.request
 
+def increment_file_name(filename, multiple):
+    filenameparts = filename.split('.')
+    return ('.'.join(filenameparts[:-1]) + " - " + str(multiple) + "." + filenameparts[1])
+
 if __name__ == "__main__":
     artistname = sys.argv[1]
     downloadfolder = "downloads"
@@ -20,8 +24,18 @@ if __name__ == "__main__":
                 
                 download = line.split('"')[1]
                 filename = download.split('/')[9]
-                urllib.request.urlretrieve(download, os.path.join(downloadfolder, artistname, filename))
-                print("Downloaded " + filename)
+                
+                if os.path.exists(os.path.join(downloadfolder, artistname, filename)):
+                    multiple = 1
+                    newfilename = increment_file_name(filename, multiple)
+                    while os.path.exists(os.path.join(downloadfolder, artistname, newfilename)):
+                        multiple += 1
+                        newfilename = increment_file_name(filename, multiple)
+                    urllib.request.urlretrieve(download, os.path.join(downloadfolder, artistname, newfilename))
+                    print("Downloaded '" + filename + "' as '" + newfilename + "'")
+                else:
+                    urllib.request.urlretrieve(download, os.path.join(downloadfolder, artistname, filename))
+                    print("Downloaded '" + filename + "'")
             elif "Show more" in line:
                 pagenum += 1
                 flag = True
